refactor(refactor): rename configfile to config
