+++ include/shared/defines.hpp
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <netinet/in.h>
+#include <stdint.h>
+#include <sys/socket.h>
+
+typedef struct epoll_event t_event;
+
+namespace www {
+
+const uint32_t LOCALHOST_ADDRESS = 0x7F000001;   // 127.0.0.1
+const uint32_t DEFAULT_MAX_BODY_SIZE = 1048576;  // 1MB
+
+}  // namespace www
