+++ src/config/Parser_keywords.cpp
@@ -1,6 +1,7 @@
 #include "config/Parser.hpp"
+#include "shared/defines.hpp"
 
-namespace configfile {
+namespace config {
 
 /**
  * @brief Converts a string to a positive integer.
@@ -54,7 +55,7 @@ int ConfigFileParser::server(std::vector<std::string>& args, int& lineCount, int
   t_server new_server;
   // set default values
   new_server.port = 80;
-  new_server.ip_address = 0x7F000001;
+  new_server.ip_address = www::LOCALHOST_ADDRESS;
   new_server.max_body_size = 1000000;
   m_configData.servers.push_back(new_server);
   // TODO: Load default Error pages if none are provided
@@ -440,4 +441,4 @@ int ConfigFileParser::index(std::vector<std::string>& args, int& lineCount, int 
   return 0;
 }
 
-} /* namespace configfile */
+}  // namespace config
